requirements.txt
certifi==2020.6.20
chainmap==1.0.3
chardet==3.0.4
combomethod==1.0.12
cycler==0.10.0
idna==2.10
joblib==0.16.0
kiwisolver==1.2.0
lxml==4.5.2
matplotlib==3.3.2
nulltype==2.3.1
numpy==1.19.2
options==1.4.10
pandas==1.1.2
pandas-datareader==0.9.0
patsy==0.5.1
Pillow==7.2.0
pipdeptree==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.1
six==1.12.0
sklearn==0.0
statsmodels==0.12.2
threadpoolctl==2.1.0
urllib3==1.25.10
cassandra-driver~=3.25.0